refactor(agent): split event types from model.go
Moved StreamEvent/PlanStep/PermissionDenial, ConvoEvent/ToolUseBlock/
ToolResultBlock/ApprovalRequest/ApprovalResponse, and ErrorEvent/
PluginErrorsEvent/EscalationEvent into event_stream.go, event_convo.go,
and event_notify.go respectively. Pure type move, no behavior change.
Reduces model.go from 1020 to ~740 lines and its conflict surface with
runner_headless.go.
