+++ internal/agent/event_stream.go
@@ -1,0 +1,64 @@
+package agent
+
+import (
+	"time"
+
+	"github.com/Automaat/sybra/internal/limits"
+)
+
+// PermissionDenial records a single auto-mode classifier denial observed during
+// a headless run. Populated from tool_result error blocks that match the Claude
+// Code auto-mode classifier denial marker.
+type PermissionDenial struct {
+	ToolUseID string
+	Reason    string
+}
+
+// PlanStep represents a single item from a TodoWrite tool call.
+type PlanStep struct {
+	Content string `json:"content"`
+	Status  string `json:"status"` // "pending", "in_progress", "completed"
+}
+
+type StreamEvent struct {
+	Type                     string  `json:"type"`
+	Content                  string  `json:"content,omitempty"`
+	SessionID                string  `json:"session_id,omitempty"`
+	CostUSD                  float64 `json:"cost_usd,omitempty"`
+	InputTokens              int     `json:"input_tokens,omitempty"`
+	OutputTokens             int     `json:"output_tokens,omitempty"`
+	CacheCreationInputTokens int     `json:"cache_creation_input_tokens,omitempty"`
+	CacheReadInputTokens     int     `json:"cache_read_input_tokens,omitempty"`
+	ReasoningTokens          int     `json:"reasoning_tokens,omitempty"`
+	// PremiumRequests is Copilot's per-result billing-unit count (result event)
+	// or 0 for claude/codex.
+	PremiumRequests float64   `json:"premium_requests,omitempty"`
+	Subtype         string    `json:"subtype,omitempty"`
+	Timestamp       time.Time `json:"timestamp"`
+	// ErrorType and ErrorStatus carry structured fields from the Anthropic error
+	// envelope (e.g. "overloaded_error", 529) when subtype == "error".
+	ErrorType   string `json:"error_type,omitempty"`
+	ErrorStatus int    `json:"error_status,omitempty"`
+	// PlanSteps is populated when the assistant calls TodoWrite; contains the
+	// latest snapshot of the agent's todo list at this point in the stream.
+	PlanSteps []PlanStep `json:"plan_steps,omitempty"`
+	// PluginErrors carries plugin load failures surfaced by the init event.
+	PluginErrors []string `json:"plugin_errors,omitempty"`
+	// ToolCalls is the number of tool_use blocks in this event: all tool uses
+	// in a Claude assistant turn, or a single Codex tool_use. The runner
+	// accumulates these into Agent.ToolCalls.
+	ToolCalls int `json:"tool_calls,omitempty"`
+	// LimitSnapshot carries provider quota status emitted by CLIs such as
+	// Codex. It is forwarded to the limits ledger and not rendered as normal
+	// assistant output.
+	LimitSnapshot *limits.Snapshot `json:"limit_snapshot,omitempty"`
+	// toolSig is a canonical fingerprint of this event's tool calls (name +
+	// input), used by the watchdog's real-time loop detector to spot an agent
+	// repeating the same call. Unexported so it is never serialized to the
+	// NDJSON log or emitted to the frontend; it lives only in memory.
+	toolSig string
+	// permissionDenials carries auto-mode classifier denial records extracted
+	// from this event's tool_result error blocks. Unexported so it is never
+	// serialized; lives in-memory only. Populated for claude "user" events only.
+	permissionDenials []PermissionDenial
+}
